Bazı düzeltmeler yapıldı.

+++ auth-service/src/controllers/role-controller/RoleController.ts
@@ -1,22 +1,23 @@
 import Role from "@/database/models/Role";
 import MyError from "@/utils/error/MyError";
 import MyErrorTypes from "@/utils/error/MyErrorTypes";
-import Logger from "@/utils/logger";
 import {
   DatabaseError,
   ForeignKeyConstraintError,
-  InferAttributes,
-  InstanceUpdateOptions,
   UniqueConstraintError,
 } from "sequelize";
 import { RoleUpdateOptions } from "./RoleControllerTypes";
 import Permission from "@/database/models/Permission";
 import PermissionRole from "@/database/models/junction_models/PermissionRole";
-import User from "@/database/models/User";
 import UserController from "../user-controller/UserController";
 import UserRole from "@/database/models/junction_models/UserRole";
 
 export default class RoleController {
+  private static staticRoleIdList: string[] = [
+    "a3cf85b7-4995-43fc-9790-58c032b27ab6",
+    "b3cf85b7-4995-43fc-9790-58c032b27ab6",
+  ];
+
   /**
    * Creates a new role with the given name and optional description.
    *
@@ -61,6 +62,10 @@ export default class RoleController {
     try {
       await Promise.all(
         ids.map(async (id) => {
+          if (this.staticRoleIdList.includes(id)) {
+            throw MyError.createError(MyErrorTypes.ROLE_DELETE_RESTRICTION);
+          }
+
           await Role.destroy({
             where: {
               id: ids,
@@ -115,6 +120,10 @@ export default class RoleController {
       if (!role) {
         throw MyError.createError(MyErrorTypes.ROLE_NOT_FOUND);
       }
+      // Check if the role name is in the restricted list.
+      if (this.staticRoleIdList.includes(role.id!)) {
+        throw MyError.createError(MyErrorTypes.ROLE_DELETE_RESTRICTION);
+      }
 
       await role.update({
         name: newData.name,
@@ -140,14 +149,15 @@ export default class RoleController {
     limit?: number
   ): Promise<{ roleList: Role[]; totalCount: number }> {
     try {
-      let offset = undefined;
+      const Paginagiton: any = {};
+
       if (limit && page) {
-        offset = (page - 1) * limit;
+        Paginagiton["offset"] = (page - 1) * limit;
+        Paginagiton["limit"] = limit;
       }
 
       const roles = await Role.findAll({
-        offset,
-        limit: limit,
+        ...Paginagiton,
       });
 
       const totalRoleCount = await Role.count();
@@ -191,6 +201,10 @@ export default class RoleController {
     permission_id: string
   ): Promise<boolean> {
     try {
+      // Check if the role name is in the restricted list.
+      if (this.staticRoleIdList.includes(role_id)) {
+        throw MyError.createError(MyErrorTypes.ROLE_DELETE_RESTRICTION);
+      }
       const deletedPermissionCount: number = await PermissionRole.destroy({
         where: {
           role_id: role_id,
@@ -217,6 +231,10 @@ export default class RoleController {
     permission_id: string
   ): Promise<boolean> {
     try {
+      // Check if the role name is in the restricted list.
+      if (this.staticRoleIdList.includes(role_id)) {
+        throw MyError.createError(MyErrorTypes.ROLE_DELETE_RESTRICTION);
+      }
       const result: PermissionRole = await PermissionRole.create({
         role_id,
         permission_id,
@@ -277,6 +295,33 @@ export default class RoleController {
           role_id: role.id,
         },
       });
+    } catch (e) {
+      throw e;
+    }
+  }
+
+  public static async GetAllPermissions(
+    page?: number,
+    limit?: number
+  ): Promise<{ permissionsList: Permission[]; totalCount: number }> {
+    try {
+      const Paginagiton: any = {};
+
+      if (limit && page) {
+        Paginagiton["offset"] = (page - 1) * limit;
+        Paginagiton["limit"] = limit;
+      }
+
+      const permissions = await Permission.findAll({
+        ...Paginagiton,
+      });
+
+      const totalPermissionCount = await Permission.count();
+      
+      return {
+        permissionsList: permissions,
+        totalCount: totalPermissionCount,
+      };
     } catch (e) {
       throw e;
     }
